backend/requirements.txt
# FitnessRPG Backend - Dépendances Python
# Installation: pip install -r requirements.txt

# Framework Flask
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0

# Base de données
SQLAlchemy==2.0.23

# Authentification
PyJWT==2.8.0
Werkzeug==3.0.1

# Validation de données
marshmallow==3.20.1

# Variables d'environnement
python-dotenv==1.0.0

# Utilitaires date/time
python-dateutil==2.8.2

# Production (optionnel)
gunicorn==21.2.0

# Développement (optionnel)
flask-migrate==4.0.5